requirements.txt
fastapi==0.115.12
mongomock==4.3.0
Pillow==11.2.1
pydantic==2.11.4
pymongo==4.12.1
pytest==8.3.5
pytest_asyncio==0.26.0
python-dotenv==1.1.0
typing_extensions==4.13.2
